+++ commands/fun/umaru.js
@@ -18,6 +18,6 @@ module.exports = {
 			let data = JSON.parse(body);	
 			let answer = data.results[0].url || "ничего найдено(";
 			message.channel.send(answer);
-		}).catch(function(e) { print_e("[ERROR/umaru.js]: Request error, "+e.message); });;
+		}).catch(function(e) { print_e("[ERROR/umaru.js] Request error, "+e.message); });;
     }
 }
