+++ commands/info/guild.js
@@ -3,6 +3,6 @@ module.exports = {
 	category: "info",
     description: "Возвращает id сервера.",
     run: async (client, message, args) => {
-        return message.channel.send("ID: `"+message.guild.id+"`").then(m => m.delete({ timeout: 10000 }));
+        return message.channel.send("ID: `"+message.guild.id+"`");
     }
 }
