+++ handlers/command.js
@@ -1,5 +1,4 @@
 const { readdirSync } = require("fs");
-
 const ascii = require("ascii-table");
 
 let table = new ascii("Commands");
